test(header): migrate a11y
